Spade Key Changes/Various Logic Fixes
Changes Spade Key on B sides to be randomized in the item pool, since there are zero repercussions for doing so unlike on the A sides. This includes removing it as a needed item in various conditions throughout the RPD, since it's realistically only useful to create shortcuts in RPD if you have other items to go around the Spade doors.

Also fixes Chief Office/Side Stair areas by flipping the region_connection for Chief Office/Side Stairs so it properly gets accounted for in UT, and also changes some logic in the surrounding area by removing the vanilla requirements to get to certain locations like the Small Storage Room. since we have teleports available. 

Also adds King Plug as a condition to access the weapon cage in the Supplies Storage Room in the sewer, to have the location properly tracked by UT.